cli: draw the meter over the ^C written on Ctrl-C
When aborting the meter (using Ctrl-C), the system draws
a "^C" string onto the terminal. This overwrites parts
of the meter display.

This re-draws the meter display so that the characters
left on the terminal after the meter has quit does not
have a "^C" string hiding potentially interesting information.