cli: Compare buffer size data_size, not pointer size
Compare the data size transferred by libusb_ctrl_transfer() to the
actual buffer size data_size, not the pointer size sizeof(data).

This happened to work on my x86_64 system, as pointer size (8)
happens to be the same size as the data size 8 which the Soundcraft
Notepad series USB communication uses.

Therefore Ambrose Li discovered this error on a Raspberry Pi with
an armv7 instruction set and 32bit memory model.

Closes: https://github.com/socratools/socradoc/issues/8