cli: Run tests on actual hardware
If actual hardware is present and is accessible (device file permissions)
then run the tests on actual hardware instead of dry-running them.

Note that this means we sequence the device access for this test suite,
and just hope no other program is messing with the device at the same
time.

Closes: https://github.com/socratools/socradoc/issues/9