cli: Run tests on actual hardware
If actual hardware is present and is accessible (device file permissions)
then run the tests on actual hardware instead of dry-running them.

This means we sequence the device access for this test suite, and just
hope no other program is messing with the device at the same time.

At the end of the test suite, we reset the ducker values to something
reasonable and turn off the ducker, but we leave the audio routing in
whatever state the last test case has left it.

Note that the dry run mode cannot be used at this time for the the
tests which would run on actual hardware if present: The scnp-cli
code is written to always detect and pass around a device, and that
cannot work in the case of no hardware being present.

Closes: https://github.com/socratools/socradoc/issues/9